[vidyadhar hudge]Add:Retrieve person belonging to city or state

+++ AddressBookUsingAdo.Net/AddressBookModel.cs
@@ -11,8 +11,8 @@ namespace AddressBookUsingAdo.Net
         public string Address { get; set; }
         public string City { get; set; }
         public string State { get; set; }
-        public int Zip { get; set; }
-        public double PhoneNumber { get; set; }
+        public Int64 Zip { get; set; }
+        public Int64 PhoneNumber { get; set; }
         public string Email { get; set; }
         public string AddressBookName { get; set; }
         public string AddressBookType { get; set; }
